Fix: 임시 저장 글 삭제 로직을 Container 로 분리
